+++ python/packaging/classic/setup.py
@@ -269,7 +269,6 @@ try:
             "pyspark.mllib.stat",
             "pyspark.ml",
             "pyspark.ml.connect",
-            "pyspark.ml.remote",
             "pyspark.ml.linalg",
             "pyspark.ml.param",
             "pyspark.ml.torch",
