+++ python/packaging/connect/setup.py
@@ -147,7 +147,6 @@ try:
         "pyspark.mllib.stat",
         "pyspark.ml",
         "pyspark.ml.connect",
-        "pyspark.ml.remote",
         "pyspark.ml.linalg",
         "pyspark.ml.param",
         "pyspark.ml.torch",
