+++ python/pyspark/ml/connect/readwrite.py
@@ -18,7 +18,7 @@
 from typing import cast, Type, TYPE_CHECKING
 
 import pyspark.sql.connect.proto as pb2
-from pyspark.ml.remote.serialize import serialize_ml_params, deserialize, deserialize_param
+from pyspark.ml.connect.serialize import serialize_ml_params, deserialize, deserialize_param
 from pyspark.ml.util import MLWriter, MLReader, RL
 from pyspark.ml.wrapper import JavaWrapper
